New versions for use with the Philemon release.
This change set provides versions of the EssenceClassLibraryExporter
that must be used with the Philemon release of the Essence# run time
system (Alpha Build 17.) They are not compatible with prior versions of
the Essence# run time system.

See the release notes for the Philemon release on the CodePlex Download
page for further details.